test: reorganize API controller tests into Feature, add AuthorizesResourceCrud concern
Split off from the budget/schedule redesign branch: moves the
Sanctum-based API controller tests from tests/Unit/Http/Controllers/API
to tests/Feature/API (matching the project's "Feature tests preferred
over Unit for HTTP" convention), extracts a shared
AuthorizesResourceCrud(ForUnverifiedUsers) concern used by several
resource CRUD tests, and carries a handful of unrelated demo/sandbox
and Dusk test fixes that were bundled into the same working branch.
