--- tests/Browser/Pages/Auth/ProfileTest.php
@@ -1,89 +1,0 @@
-<?php
-
-namespace Tests\Browser\Pages\Auth;
-
-use App\Models\User;
-use Illuminate\Foundation\Testing\DatabaseMigrations;
-use Laravel\Dusk\Browser;
-use PHPUnit\Framework\Attributes\Group;
-use Tests\DuskTestCase;
-
-#[Group('extended')]
-class ProfileTest extends DuskTestCase
-{
-    use DatabaseMigrations;
-
-    public function test_password_change_client_side_validations(): void
-    {
-        // Make sure the sandbox mode is disabled
-        // This ensures that the client is allowed to change the password
-        $this->setConfig('yaffa.sandbox_mode', false);
-
-        /** @var User $user */
-        $user = User::factory()->create([
-            'language' => 'en'
-        ]);
-
-        $this->browse(function (Browser $browser) use ($user) {
-            $browser->loginAs($user)
-                ->visit('/user/settings')
-                ->waitFor('#changePasswordForm')
-
-                // Validate that no errors are shown initially in the change password form
-                ->assertMissing('#changePasswordForm div.invalid-feedback')
-
-                // Type incorrect current password
-                ->type('current_password', 'incorrect')
-
-                // Type new password that is too short
-                ->type('password', 'short')
-
-                // Type new password that does not match the confirmation
-                ->type('password_confirmation', 'doesnotmatch')
-
-                // Submit the form and wait for the validation errors, including the toast message
-                ->press('@button-change-password')
-                ->waitForTextIn('div.toast-container div.toast.bg-danger.show', 'Validation failed. Please check the form for errors.', 30)
-                ->waitForTextIn('#current_password + div.invalid-feedback', 'The password is incorrect.')
-                ->waitForTextIn('#password + div.invalid-feedback', 'The password must be at least 8 characters.')
-
-                // Set the correct password, a valid new password with a non-matching confirmation
-                ->waitUntilMissing('div.toast-container div.toast.bg-danger.show', 30)
-                ->type('current_password', 'password')
-                ->type('password', 'password123')
-                ->type('password_confirmation', 'password1234')
-
-                // Submit the form and wait for the validation errors, including the toast message
-                ->press('@button-change-password')
-                ->waitForTextIn('div.toast-container div.toast.bg-danger.show', 'Validation failed. Please check the form for errors.', 30)
-                ->waitForTextIn('#password + div.invalid-feedback', 'The password confirmation does not match.')
-
-                // Set the correct password, a valid new password with a matching confirmation
-                ->waitUntilMissing('div.toast-container div.toast.bg-danger.show', 30)
-                ->type('current_password', 'password')
-                ->type('password', 'password123')
-                ->type('password_confirmation', 'password123')
-
-                // Submit the form and wait for the success toast message
-                ->press('@button-change-password')
-                ->waitForTextIn('div.toast-container div.toast.bg-success.show', 'Password changed successfully.')
-
-                // The form should be reset
-                ->assertInputValue('current_password', '')
-                ->assertInputValue('password', '')
-                ->assertInputValue('password_confirmation', '');
-
-            // Now log out and log back in with the new password
-            $browser
-                ->logout()
-                ->visitRoute('home')
-                // Page should redirect to login
-                ->waitForRoute('login', [], 10)
-                ->type('email', $user->email)
-                ->type('password', 'password123')
-                ->press('@login-button')
-                ->waitForRoute('home', [], 10)
-                ->assertRouteIs('home');
-        });
-    }
-}
