+++ tests/Browser/Pages/Transactions/TransactionFormInvestmentModalTest.php
@@ -3,8 +3,6 @@
 namespace Tests\Browser\Pages\Transactions;
 
 use App\Models\AccountEntity;
-use App\Models\Investment;
-use App\Models\InvestmentPrice;
 use App\Models\Transaction;
 use App\Models\User;
 use Laravel\Dusk\Browser;
@@ -204,188 +202,4 @@ class TransactionFormInvestmentModalTest extends DuskTestCase
         });
     }
 
-    public function test_store_price_checkbox_is_visible_when_no_existing_price(): void
-    {
-        $this->browse(function (Browser $browser) {
-            $browser->loginAs($this->user)
-                // Load the view for a random account
-                ->visitRoute('account-entity.show', ['account_entity' => $this->accountEntity->id])
-                // Wait for the page to load
-                ->waitForText('Account details')
-                // Click the "new investment transaction" button
-                ->click('#create-investment-transaction-button')
-                // Wait for the modal to load
-                ->waitForText('Finalize transaction draft')
-                ->waitFor('#transactionFormInvestment')
-                ->waitFor('#account', 10)
-                ->waitFor('#investment', 10)
-
-                // Select investment
-                ->select2('#investment', null, 10)
-                // Select transaction type
-                ->select('#transaction_type', 'buy')
-                // Add quantity
-                ->type('#transaction_quantity', '10')
-                // Add price - this should trigger the price check
-                ->type('#transaction_price', '25.50')
-                // Wait for the store price checkbox to appear
-                ->waitFor('#store_price_checkbox', 10)
-                // Verify the checkbox is visible
-                ->assertVisible('#store_price_checkbox')
-                ->assertVisible('label[for="store_price_checkbox"]');
-        });
-    }
-
-    public function test_store_price_checkbox_is_not_visible_when_price_exists(): void
-    {
-        // First create an investment price for today
-        $investment = Investment::where('name', 'Test investment USD')->first();
-        InvestmentPrice::create([
-            'investment_id' => $investment->id,
-            'date' => now()->format('Y-m-d'),
-            'price' => 100.00,
-        ]);
-
-        $this->browse(function (Browser $browser) use ($investment) {
-            $browser->loginAs($this->user)
-                // Load the view for a random account
-                ->visitRoute('account-entity.show', ['account_entity' => $this->accountEntity->id])
-                // Wait for the page to load
-                ->waitForText('Account details')
-                // Click the "new investment transaction" button
-                ->click('#create-investment-transaction-button')
-                // Wait for the modal to load
-                ->waitForText('Finalize transaction draft')
-                ->waitFor('#transactionFormInvestment')
-                ->waitFor('#account', 10)
-                ->waitFor('#investment', 10)
-
-                // Select transaction type - although generally Buy is the default selection
-                ->select('#transaction_type', 'buy')
-                // Select the specific investment that has a price - this should trigger the price check
-                ->select2ExactSearch('#investment', $investment->name, 10)
-                // Wait a bit for the API call to complete, the price info should be visible now
-                ->waitFor('span.existing-price-label', 10)
-                // Verify the store price checkbox is NOT visible
-                ->assertMissing('#store_price_checkbox');
-        });
-    }
-
-    public function test_user_can_save_transaction_with_price_storage_enabled(): void
-    {
-        // Get an investment without price for today
-        $investment = Investment::where('name', 'Test investment USD')->first();
-
-        // Ensure no price exists for today
-        InvestmentPrice::where('investment_id', $investment->id)
-            ->where('date', now()->format('Y-m-d'))
-            ->delete();
-
-        $this->browse(function (Browser $browser) use ($investment) {
-            $browser->loginAs($this->user)
-                // Load the view for a random account
-                ->visitRoute('account-entity.show', ['account_entity' => $this->accountEntity->id])
-                // Wait for the page to load
-                ->waitForText('Account details')
-                // Click the "new investment transaction" button
-                ->click('#create-investment-transaction-button')
-                // Wait for the modal to load
-                ->waitForText('Finalize transaction draft')
-                ->waitFor('#transactionFormInvestment')
-                ->waitFor('#account', 10)
-                ->waitFor('#investment', 10)
-
-                // Fill the form
-                ->select2ExactSearch('#investment', $investment->name, 10)
-                ->select('#transaction_type', 'buy')
-                ->type('#transaction_quantity', '10')
-                ->type('#transaction_price', '35.75')
-                ->type('#transaction_commission', '5')
-                ->type('#transaction_tax', '2')
-
-                // Wait for store price checkbox to appear and enable it
-                ->waitFor('#store_price_checkbox', 10)
-                ->click('label[for="store_price_checkbox"]')
-                ->assertChecked('#store_price_checkbox')
-
-                // Submit form
-                ->click('#transactionFormInvestment-Save')
-                // Two success messages should be available
-                ->waitFor('.toast-container .toast.bg-success.show', 10)
-                ->pause(500)  // Give both toasts time to render
-                ->assertSeeIn('.toast-container', 'Transaction added')
-                ->assertSeeIn('.toast-container', 'Investment price stored')
-                // Wait for the modal to close
-                ->waitUntilMissing('#modal-transaction-form-investment', 10);
-
-            // Verify the transaction was saved in the database
-            $transaction = Transaction::orderByDesc('id')->first();
-            $this->assertNotNull($transaction);
-            $this->assertEquals(10, $transaction->config->quantity->toFloat());
-            $this->assertEquals(35.75, $transaction->config->price->getAmount()->toFloat());
-
-            // Verify the investment price was saved in the database
-            $investmentPrice = InvestmentPrice::where('investment_id', $investment->id)
-                ->where('date', now()->format('Y-m-d'))
-                ->first();
-            $this->assertNotNull($investmentPrice);
-            $this->assertEquals(35.75, $investmentPrice->price->getAmount()->toFloat());
-        });
-    }
-
-    public function test_user_can_save_transaction_without_price_storage(): void
-    {
-        // Get an investment without price for today
-        $investment = Investment::where('name', 'Test investment USD')->first();
-
-        // Ensure no price exists for today
-        InvestmentPrice::where('investment_id', $investment->id)
-            ->where('date', now()->format('Y-m-d'))
-            ->delete();
-
-        $this->browse(function (Browser $browser) use ($investment) {
-            $browser->loginAs($this->user)
-                // Load the view for an account with the same currency as the investment
-                ->visitRoute('account-entity.show', ['account_entity' => $this->accountEntity->id])
-                // Wait for the page to load
-                ->waitForText('Account details')
-                // Click the "new investment transaction" button
-                ->click('#create-investment-transaction-button')
-                // Wait for the modal to load
-                ->waitForText('Finalize transaction draft')
-                ->waitFor('#transactionFormInvestment')
-                ->waitFor('#account', 10)
-                ->waitFor('#investment', 10)
-
-                // Fill the form
-                ->select2ExactSearch('#investment', $investment->name, 10)
-                ->select('#transaction_type', 'buy')
-                ->type('#transaction_quantity', '5')
-                ->type('#transaction_price', '42.25')
-                ->type('#transaction_commission', '3')
-
-                // Wait for store price checkbox to appear but DO NOT enable it
-                ->waitFor('#store_price_checkbox', 10)
-                ->assertNotChecked('#store_price_checkbox')
-
-                // Submit form
-                ->click('#transactionFormInvestment-Save')
-                // Wait for the modal to close
-                ->waitUntilMissing('#modal-transaction-form-investment', 10)
-                // A success message should be available for the transaction
-                ->waitForTextIn('.toast-container .toast.bg-success.show', 'Transaction added', 10);
-
-            // Verify the transaction was saved
-            $transaction = Transaction::orderByDesc('id')->first();
-            $this->assertNotNull($transaction);
-            $this->assertEquals(5, $transaction->config->quantity->toFloat());
-            $this->assertEquals(42.25, $transaction->config->price->getAmount()->toFloat());
-
-            // Verify the investment price was NOT saved in the database
-            $investmentPrice = InvestmentPrice::where('investment_id', $investment->id)
-                ->where('date', now()->format('Y-m-d'))
-                ->first();
-            $this->assertNull($investmentPrice);
-        });
-    }
 }
