+++ tests/Feature/API/V1/InvestmentPriceApiV1Test.php
@@ -59,7 +59,9 @@ class InvestmentPriceApiV1Test extends TestCase
 
     public function test_v1_index_returns_prices(): void
     {
-        InvestmentPrice::factory()->for($this->investment)->count(3)->create();
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-01', 'price' => 100]);
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-02', 'price' => 110]);
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-03', 'price' => 120]);
 
         $response = $this->actingAs($this->user)
             ->getJson(route('api.v1.investment-prices.index', ['investment' => $this->investment->id]));
@@ -71,6 +73,48 @@ class InvestmentPriceApiV1Test extends TestCase
                 ],
             ])
             ->assertJsonCount(3, 'prices');
+
+        // Ordered by date.
+        $json = $response->json();
+        $this->assertSame('2024-01-01', $json['prices'][0]['date']);
+        $this->assertSame('2024-01-02', $json['prices'][1]['date']);
+        $this->assertSame('2024-01-03', $json['prices'][2]['date']);
+    }
+
+    public function test_v1_index_filters_by_date_from(): void
+    {
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-01', 'price' => 100]);
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-15', 'price' => 110]);
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-30', 'price' => 120]);
+
+        $response = $this->actingAs($this->user)
+            ->getJson(route('api.v1.investment-prices.index', [
+                'investment' => $this->investment->id,
+                'date_from' => '2024-01-15',
+            ]));
+
+        $response->assertOk()->assertJsonCount(2, 'prices');
+        $json = $response->json();
+        $this->assertSame('2024-01-15', $json['prices'][0]['date']);
+        $this->assertSame('2024-01-30', $json['prices'][1]['date']);
+    }
+
+    public function test_v1_index_filters_by_date_to(): void
+    {
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-01', 'price' => 100]);
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-15', 'price' => 110]);
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-30', 'price' => 120]);
+
+        $response = $this->actingAs($this->user)
+            ->getJson(route('api.v1.investment-prices.index', [
+                'investment' => $this->investment->id,
+                'date_to' => '2024-01-15',
+            ]));
+
+        $response->assertOk()->assertJsonCount(2, 'prices');
+        $json = $response->json();
+        $this->assertSame('2024-01-01', $json['prices'][0]['date']);
+        $this->assertSame('2024-01-15', $json['prices'][1]['date']);
     }
 
     public function test_v1_index_filters_by_date_range(): void
@@ -102,14 +146,95 @@ class InvestmentPriceApiV1Test extends TestCase
 
         $response->assertCreated()
             ->assertJsonStructure([
-                'price' => ['id', 'investment_id', 'date', 'price'],
+                'price' => ['id', 'investment_id', 'date', 'price', 'created_at', 'updated_at', 'investment'],
                 'message',
-            ]);
+            ])
+            ->assertJsonPath('price.date', '2024-01-15')
+            ->assertJsonPath('price.price', '100.5000000000')
+            ->assertJsonPath('price.investment_id', $this->investment->id);
 
         $this->assertDatabaseHas('investment_prices', [
             'investment_id' => $data['investment_id'],
             'date' => $data['date'],
         ]);
+    }
+
+    public function test_v1_store_requires_investment_id(): void
+    {
+        $response = $this->actingAs($this->user)
+            ->postJson(route('api.v1.investment-prices.store'), [
+                'date' => '2024-01-15',
+                'price' => 150.50,
+            ]);
+
+        $response->assertUnprocessable()->assertJsonValidationErrors(['investment_id']);
+    }
+
+    public function test_v1_store_requires_date(): void
+    {
+        $response = $this->actingAs($this->user)
+            ->postJson(route('api.v1.investment-prices.store'), [
+                'investment_id' => $this->investment->id,
+                'price' => 150.50,
+            ]);
+
+        $response->assertUnprocessable()->assertJsonValidationErrors(['date']);
+    }
+
+    public function test_v1_store_requires_price(): void
+    {
+        $response = $this->actingAs($this->user)
+            ->postJson(route('api.v1.investment-prices.store'), [
+                'investment_id' => $this->investment->id,
+                'date' => '2024-01-15',
+            ]);
+
+        $response->assertUnprocessable()->assertJsonValidationErrors(['price']);
+    }
+
+    public function test_v1_store_prevents_duplicate_prices_for_same_date(): void
+    {
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-15', 'price' => 100]);
+
+        $response = $this->actingAs($this->user)
+            ->postJson(route('api.v1.investment-prices.store'), [
+                'investment_id' => $this->investment->id,
+                'date' => '2024-01-15',
+                'price' => 200,
+            ]);
+
+        $response->assertUnprocessable()->assertJsonValidationErrors(['date']);
+    }
+
+    public function test_v1_price_must_be_numeric(): void
+    {
+        $response = $this->actingAs($this->user)
+            ->postJson(route('api.v1.investment-prices.store'), [
+                'investment_id' => $this->investment->id,
+                'date' => '2024-01-15',
+                'price' => 'not-a-number',
+            ]);
+
+        $response->assertUnprocessable()->assertJsonValidationErrors(['price']);
+    }
+
+    public function test_v1_price_must_be_greater_than_zero(): void
+    {
+        $response = $this->actingAs($this->user)
+            ->postJson(route('api.v1.investment-prices.store'), [
+                'investment_id' => $this->investment->id,
+                'date' => '2024-01-15',
+                'price' => 0,
+            ]);
+        $response->assertUnprocessable()->assertJsonValidationErrors(['price']);
+
+        $response = $this->actingAs($this->user)
+            ->postJson(route('api.v1.investment-prices.store'), [
+                'investment_id' => $this->investment->id,
+                'date' => '2024-01-15',
+                'price' => -100,
+            ]);
+        $response->assertUnprocessable()->assertJsonValidationErrors(['price']);
     }
 
     public function test_v1_update_modifies_price(): void
@@ -129,12 +254,49 @@ class InvestmentPriceApiV1Test extends TestCase
             ->assertJsonStructure([
                 'price' => ['id', 'investment_id', 'date', 'price'],
                 'message',
-            ]);
+            ])
+            ->assertJsonPath('price.date', '2024-01-16')
+            ->assertJsonPath('price.price', '110.0000000000')
+            ->assertJsonPath('message', __('Investment price updated'));
 
         $this->assertDatabaseHas('investment_prices', [
             'id' => $price->id,
             'price' => 110.00,
         ]);
+    }
+
+    public function test_v1_update_cross_user_investment_id_fails_validation(): void
+    {
+        // A PUT with an investment_id the acting user doesn't own fails validation (422), not
+        // authorization (403) - investment_id ownership is enforced by the request's validation
+        // rule, not a policy check, so this is a deliberately different status than the
+        // GET/DELETE cross-user cases below.
+        $otherUser = User::factory()->create();
+        $this->createForUser($otherUser, Currency::class);
+        $otherInvestment = Investment::factory()->for($otherUser)->withUser($otherUser)->create();
+        $price = InvestmentPrice::factory()->for($otherInvestment)->create(['date' => '2024-01-01', 'price' => 100]);
+
+        $response = $this->actingAs($this->user)
+            ->putJson(route('api.v1.investment-prices.update', ['investmentPrice' => $price->id]), [
+                'investment_id' => $otherInvestment->id,
+                'date' => '2024-01-02',
+                'price' => 200,
+            ]);
+
+        $response->assertUnprocessable()->assertJsonValidationErrors(['investment_id']);
+    }
+
+    public function test_v1_destroy_cross_user_price_is_forbidden(): void
+    {
+        $otherUser = User::factory()->create();
+        $this->createForUser($otherUser, Currency::class);
+        $otherInvestment = Investment::factory()->for($otherUser)->withUser($otherUser)->create();
+        $price = InvestmentPrice::factory()->for($otherInvestment)->create(['date' => '2024-01-01', 'price' => 100]);
+
+        $response = $this->actingAs($this->user)
+            ->deleteJson(route('api.v1.investment-prices.destroy', ['investmentPrice' => $price->id]));
+
+        $response->assertForbidden();
     }
 
     public function test_v1_destroy_deletes_price(): void
@@ -144,9 +306,27 @@ class InvestmentPriceApiV1Test extends TestCase
         $response = $this->actingAs($this->user)
             ->deleteJson(route('api.v1.investment-prices.destroy', ['investmentPrice' => $price->id]));
 
-        $response->assertOk()->assertJsonStructure(['message']);
+        $response->assertOk()
+            ->assertJsonStructure(['message'])
+            ->assertJsonPath('message', __('Investment price deleted'));
 
         $this->assertDatabaseMissing('investment_prices', ['id' => $price->id]);
+    }
+
+    public function test_v1_check_price_validates_date_format(): void
+    {
+        // Missing date parameter
+        $response = $this->actingAs($this->user)
+            ->getJson(route('api.v1.investment-prices.check', ['investment' => $this->investment->id]));
+        $response->assertUnprocessable()->assertJsonValidationErrors(['date']);
+
+        // Invalid date formats
+        foreach (['01/15/2024', '2024-13-01', 'not-a-date'] as $invalidDate) {
+            $response = $this->actingAs($this->user)->getJson(
+                route('api.v1.investment-prices.check', ['investment' => $this->investment->id, 'date' => $invalidDate])
+            );
+            $response->assertUnprocessable()->assertJsonValidationErrors(['date']);
+        }
     }
 
     public function test_v1_check_price_returns_exists_false_when_no_price(): void
@@ -166,7 +346,24 @@ class InvestmentPriceApiV1Test extends TestCase
             ->getJson(route('api.v1.investment-prices.check', ['investment' => $this->investment->id]) . '?date=2024-01-15');
 
         $response->assertOk()
-            ->assertJson(['exists' => true]);
+            ->assertJson(['exists' => true])
+            ->assertJsonPath('price', '99.9900000000');
+    }
+
+    public function test_v1_check_price_only_checks_specific_investment(): void
+    {
+        $otherInvestment = Investment::factory()->for($this->user)->withUser($this->user)->create();
+        InvestmentPrice::factory()->for($this->investment)->create(['date' => '2024-01-15', 'price' => 123.45]);
+
+        $response = $this->actingAs($this->user)->getJson(
+            route('api.v1.investment-prices.check', ['investment' => $this->investment->id]) . '?date=2024-01-15'
+        );
+        $response->assertOk()->assertJson(['exists' => true]);
+
+        $response = $this->actingAs($this->user)->getJson(
+            route('api.v1.investment-prices.check', ['investment' => $otherInvestment->id]) . '?date=2024-01-15'
+        );
+        $response->assertOk()->assertJson(['exists' => false]);
     }
 
     // ===== ERROR FORMAT TESTS =====
@@ -177,15 +374,7 @@ class InvestmentPriceApiV1Test extends TestCase
             ->postJson(route('api.v1.investment-prices.store'), []);
 
         $response->assertUnprocessable()
-            ->assertJsonStructure(['message', 'errors' => ['investment_id', 'date', 'price']]);
-    }
-
-    public function test_v1_validation_error_includes_field_details(): void
-    {
-        $response = $this->actingAs($this->user)
-            ->postJson(route('api.v1.investment-prices.store'), []);
-
-        $response->assertUnprocessable()
+            ->assertJsonStructure(['message', 'errors' => ['investment_id', 'date', 'price']])
             ->assertJsonValidationErrors(['investment_id', 'date', 'price']);
     }
