+++ tests/Feature/AccountTest.php
@@ -13,10 +13,12 @@ use App\Models\Transaction;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\Response;
+use Tests\Feature\Concerns\AuthorizesResourceCrud;
 use Tests\TestCase;
 
 class AccountTest extends TestCase
 {
+    use AuthorizesResourceCrud;
     use RefreshDatabase;
 
     private function createAccountAndUser(): AccountEntity
@@ -44,6 +46,27 @@ class AccountTest extends TestCase
         $this->setBaseModel(AccountEntity::class);
     }
 
+    protected function resourceAuthCollectionRouteParams(): array
+    {
+        return ['type' => 'account'];
+    }
+
+    protected function resourceAuthMemberRouteParams($resource): array
+    {
+        return ['type' => 'account', 'account_entity' => $resource->id];
+    }
+
+    protected function resourceAuthSupportsDestroy(): bool
+    {
+        // account-entity.destroy does not exist.
+        return false;
+    }
+
+    protected function createResourceForAuthTest(User $user): AccountEntity
+    {
+        return $this->createAccountAndUser();
+    }
+
     public function test_user_cannot_create_new_account_without_an_account_group(): void
     {
         /** @var User $user */
@@ -68,41 +91,6 @@ class AccountTest extends TestCase
 
         // User is redirected to create a currency first
         $response->assertRedirectToRoute('currencies.create');
-    }
-
-    public function test_guest_cannot_access_resource(): void
-    {
-        // Unauthenticated user cannot access any actions of the resource
-        $this->get(route("{$this->base_route}.index", ['type' => 'account']))->assertRedirectToRoute('login');
-        $this->get(route("{$this->base_route}.create", ['type' => 'account']))->assertRedirectToRoute('login');
-        $this->post(route("{$this->base_route}.store", ['type' => 'account']))->assertRedirectToRoute('login');
-
-        // Create a user and the related resources
-        $account = $this->createAccountAndUser();
-
-        $this->get(route("{$this->base_route}.edit", ['type' => 'account', 'account_entity' => $account->id]))
-            ->assertRedirectToRoute('login');
-        $this->patch(route("{$this->base_route}.update", ['type' => 'account', 'account_entity' => $account->id]))
-            ->assertRedirectToRoute('login');
-    }
-
-    public function test_user_cannot_access_other_users_resource(): void
-    {
-        $account = $this->createAccountAndUser();
-
-        /** @var User $user2 */
-        $user2 = User::factory()->create();
-
-        $this->actingAs($user2)->get(route("{$this->base_route}.edit", [
-            'type' => 'account',
-            'account_entity' => $account->id
-        ]))
-            ->assertStatus(Response::HTTP_FORBIDDEN);
-        $this->actingAs($user2)->patch(route("{$this->base_route}.update", [
-            'type' => 'account',
-            'account_entity' => $account->id
-        ]))
-            ->assertStatus(Response::HTTP_FORBIDDEN);
     }
 
     public function test_user_cannot_view_other_users_account_history(): void
