+++ tests/Feature/PayeeTest.php
@@ -8,10 +8,12 @@ use App\Models\Payee;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\Response;
+use Tests\Feature\Concerns\AuthorizesResourceCrud;
 use Tests\TestCase;
 
 class PayeeTest extends TestCase
 {
+    use AuthorizesResourceCrud;
     use RefreshDatabase;
 
     protected function setUp(): void
@@ -22,38 +24,25 @@ class PayeeTest extends TestCase
         $this->setBaseModel(AccountEntity::class);
     }
 
-    public function test_guest_cannot_access_resource(): void
+    protected function resourceAuthCollectionRouteParams(): array
     {
-        $this->get(route("{$this->base_route}.index", ['type' => 'payee']))->assertRedirectToRoute('login');
-        $this->get(route("{$this->base_route}.create", ['type' => 'payee']))->assertRedirectToRoute('login');
-        $this->post(route("{$this->base_route}.store", ['type' => 'payee']))->assertRedirectToRoute('login');
-
-        /** @var User $user */
-        $user = User::factory()->create();
-        Category::factory()->for($user)->create();
-        /** @var AccountEntity $payee */
-        $payee = AccountEntity::factory()->for($user)->for(Payee::factory()->withUser($user), 'config')->create();
-
-        $this->get(route("{$this->base_route}.edit", ['account_entity' => $payee->id]))
-            ->assertRedirectToRoute('login');
-        $this->patch(route("{$this->base_route}.update", ['account_entity' => $payee->id]))
-            ->assertRedirectToRoute('login');
+        return ['type' => 'payee'];
     }
 
-    public function test_user_cannot_access_other_users_resource(): void
+    protected function resourceAuthMemberRouteParams($resource): array
     {
-        /** @var User $user1 */
-        $user1 = User::factory()->create();
+        return ['account_entity' => $resource->id];
+    }
 
-        /** @var AccountEntity $payee */
-        $payee = AccountEntity::factory()->for($user1)->for(Payee::factory()->withUser($user1), 'config')->create();
+    protected function resourceAuthSupportsDestroy(): bool
+    {
+        // account-entity.destroy does not exist - payees are only removed via the merge flow.
+        return false;
+    }
 
-        /** @var User $user2 */
-        $user2 = User::factory()->create();
-        $this->actingAs($user2)->get(route("{$this->base_route}.edit", ['account_entity' => $payee->id]))
-            ->assertStatus(Response::HTTP_FORBIDDEN);
-        $this->actingAs($user2)->patch(route("{$this->base_route}.update", ['account_entity' => $payee->id]))
-            ->assertStatus(Response::HTTP_FORBIDDEN);
+    protected function createResourceForAuthTest(User $user): AccountEntity
+    {
+        return AccountEntity::factory()->for($user)->for(Payee::factory()->withUser($user), 'config')->create();
     }
 
     public function test_user_can_view_list_of_payees(): void
